renamed components to shared-components. Updated imports.
