+++ src/App.jsx
@@ -1,11 +1,12 @@
 import { BrowserRouter, Routes, Route } from "react-router-dom";
-import Nav from "./components/Nav";
-import Profile from "./components/Profile";
+import Nav from "./shared-components/Nav";
+import Profile from "./shared-components/Profile";
 import Home from "./pages/Home";
 import Login from "./pages/Login";
 import Search from "./pages/Search";
 import NewRecipe from "./pages/NewRecipe";
 import Recipe from "./pages/Recipe";
+import Ingredients from "./pages/Ingredients";
 
 function App() {
   return (
