+++ src/pages/ingredients.css
@@ -1,8 +1,6 @@
 .container {
   width: 1200px;
   height: 100%;
-  margin-left: 30%;
-  margin-right: 30%;
 }
 
 .container {
@@ -17,6 +15,7 @@
   align-items: center;
   margin-top: 10%;
 }
+
 .two h1 {
   font-size: 5rem;
   font-family: Inter;
@@ -90,6 +89,7 @@ main {
   background-repeat: no-repeat;
   background-size: cover;
 }
+
 .list li {
   list-style-type: none;
 }
@@ -125,6 +125,7 @@ main {
   flex-direction: column;
   justify-self: space-evenly;
 }
+
 .five .buttons {
   display: flex;
   justify-content: row;
