+++ src/shared-components/CommonIngredientsDisplay.jsx
