+++ src/shared-components/DietaryTags.jsx
