+++ src/shared-components/profile.css
@@ -1,8 +1,6 @@
 .container {
   width: 1200px;
   height: 100%;
-  margin-left: 30%;
-  margin-right: 30%;
 }
 
 .container {
@@ -17,6 +15,7 @@
   align-items: center;
   margin-top: 10%;
 }
+
 .two h1 {
   font-size: 5rem;
   font-family: Inter;
@@ -80,6 +79,7 @@ main {
   background-repeat: no-repeat;
   background-size: cover;
 }
+
 .list li {
   list-style-type: none;
 }
